move to `cli` dir, refactor for multiple subcommands
