+++ .travis.yml
@@ -66,7 +66,7 @@ matrix:
         - cargo doc -v
         - cargo test -v
 
-    - name: Publisher CLI
+    - name: Fluence CLI
 
       language: rust
       rust: stable
@@ -74,7 +74,7 @@ matrix:
 
       before_install:
         - rustup component add rustfmt-preview
-        - cd publisher
+        - cd cli
 
       script:
         - npm install --prefix ../bootstrap
