+++ cli/scripts/swarm-simulation.js
